Python/Tasks/thawe-marine-service/requirement.txt
aniso8601==8.0.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
cheroot==8.2.1
CherryPy==18.5.0
Click==7.0
Flask==1.1.1
Flask-Log-Request-ID==0.10.1
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
more-itertools==8.1.0
portend==2.6
pygelf==0.3.6
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
tempora==2.1.0
timeloop==1.0.2
urllib3==1.25.7
Werkzeug==0.16.0
zc.lockfile==2.0
